Move dashboard and answer, add dashboard default text
