+++ src/dashboard.js
@@ -1,4 +1,5 @@
-const { NullAttributeError } = require('../../error')
+const { NullAttributeError } = require('./error')
+const LOCALE = require('./locale');
 
 class Dashboard {
     constructor(idea) {
@@ -12,7 +13,8 @@ class Dashboard {
     toMarkdown() {
         var markdown = "### Dashboard\n\n";
 
-        //TODO
+        // Add dashboard
+        markdown += LOCALE.GITHUB.get("github.dashboard");
 
         return markdown;
     }
