+++ static/properties/github.properties
@@ -1,5 +1,6 @@
 github.private = > **Note**\n> This repository is private, you and your collaborators only can see this business idea.\n\n
 github.public = > **Warning**\n> This repository is public, anybody can see this business idea. Go to [settings]({0}) to change the visibility.\n\n
+github.dashboard = _Your business idea dashboard will soon be available here with:_\n- _A resume of the idea_\n- _A progress bar to visually represent the progress of your idea_\n- _A score to understand how successful your business might be_\n\n_Something is missing? [Make a propsal](https://github.com/stivbot/stivbot)_
 
 github.new.1 = Hello, my name is Stiv. My goal is to assist you to enhance your business ideas.\n\nDo you want my help on this issue?\n\nReact by 👍 if you need my help and 👎 if you don't.
